+++ System/core/Object/LongNote/LongNote.hpp
@@ -20,7 +20,7 @@ namespace ct
         LongNote(int type, double firstCount, double speed, std::shared_ptr<Note>& parent);
         virtual ~LongNote() = default;
         virtual bool update(double nowCount, double countPerFrame)override;
-        virtual void diffDraw(double count, float scrollRate)const override;
+        virtual void diffDraw(double count, double scrollRate)const override;
         const double& getSpeed()const { return m_scrollSpeed; }
         const  NoteType getType()const { return m_type; }
