+++ System/core/Object/Note/Note.hpp
@@ -34,7 +34,7 @@ namespace ct
         virtual ~Note() = default;
         virtual void init()override;
         virtual bool update(double nowCount, double countPerFrame) override;
-        virtual void diffDraw(double count, float scrollRate)const override;
+        virtual void diffDraw(double count, double scrollRate)const override;
         const s3d::Color& getColor()const { return m_color; }
         const  NoteType getType()const { return m_type; }
         const double& getSpeed()const { return m_scrollSpeed; }
